Moved code from CheckoutConfirmation to Cart
CheckoutConfirmation was basically used as a test for cart cards.
